+++ modules/errfinder.pl
@@ -1,7 +1,7 @@
-#start Checking error.log
+#start Checking common logs
 dprint("Finding common log files name");
 $ertf=0;
-@error = ('error.log','error_log','php-scripts.log','php.errors','php5-fpm.log','php_errors.log','debug.log');
+@error = ('error.log','error_log','php-scripts.log','php.errors','php5-fpm.log','php_errors.log','debug.log','security.txt','.well-known/security.txt');
 foreach $er(@error){
     if (($content_type, $doc_length, $mod_time, $expires, $server) =head("$target/$er")){
         if($content_type !~ m/text\/html/i){
@@ -13,4 +13,4 @@ foreach $er(@error){
 if($ertf==0) {
     fprint("error log is not found");
 }
-#end Checking error.log
+#end Checking common logs